moved around commands with new limits in mind
