+++ src/main/java/de/kittybot/kittybot/commands/info/DashboardCommand.java
@@ -1,5 +1,7 @@
-package de.kittybot.kittybot.commands.info.info;
+package de.kittybot.kittybot.commands.info;
 
+import de.kittybot.kittybot.slashcommands.application.Category;
+import de.kittybot.kittybot.slashcommands.application.RunCommand;
 import de.kittybot.kittybot.slashcommands.application.options.SubCommand;
 import de.kittybot.kittybot.slashcommands.interaction.Interaction;
 import de.kittybot.kittybot.slashcommands.interaction.Options;
@@ -7,10 +9,10 @@ import de.kittybot.kittybot.utils.Config;
 import de.kittybot.kittybot.utils.MessageUtils;
 
 @SuppressWarnings("unused")
-public class DashboardCommand extends SubCommand{
+public class DashboardCommand extends RunCommand{
 
 	public DashboardCommand(){
-		super("dashboard", "Shows you our dashboard");
+		super("dashboard", "Shows you our dashboard", Category.INFORMATION);
 	}
 
 	@Override
