+++ src/main/java/de/kittybot/kittybot/commands/info/PingCommand.java
@@ -1,15 +1,17 @@
-package de.kittybot.kittybot.commands.info.info;
+package de.kittybot.kittybot.commands.info;
 
+import de.kittybot.kittybot.slashcommands.application.Category;
+import de.kittybot.kittybot.slashcommands.application.RunCommand;
 import de.kittybot.kittybot.slashcommands.application.options.SubCommand;
 import de.kittybot.kittybot.slashcommands.interaction.Interaction;
 import de.kittybot.kittybot.slashcommands.interaction.Options;
 import de.kittybot.kittybot.utils.Config;
 
 @SuppressWarnings("unused")
-public class PingCommand extends SubCommand{
+public class PingCommand extends RunCommand{
 
 	public PingCommand(){
-		super("ping", "Shows the bots ping");
+		super("ping", "Shows the bots ping", Category.INFORMATION);
 	}
 
 	@Override
