+++ src/main/java/de/kittybot/kittybot/commands/info/PrivacyCommand.java
@@ -1,5 +1,7 @@
-package de.kittybot.kittybot.commands.info.info;
+package de.kittybot.kittybot.commands.info;
 
+import de.kittybot.kittybot.slashcommands.application.Category;
+import de.kittybot.kittybot.slashcommands.application.RunCommand;
 import de.kittybot.kittybot.slashcommands.application.options.SubCommand;
 import de.kittybot.kittybot.slashcommands.interaction.Interaction;
 import de.kittybot.kittybot.slashcommands.interaction.Options;
@@ -7,10 +9,10 @@ import de.kittybot.kittybot.utils.Config;
 import de.kittybot.kittybot.utils.MessageUtils;
 
 @SuppressWarnings("unused")
-public class PrivacyCommand extends SubCommand{
+public class PrivacyCommand extends RunCommand{
 
 	public PrivacyCommand(){
-		super("privacy", "Gives you a link to our privacy policy");
+		super("privacy", "Gives you a link to our privacy policy", Category.INFORMATION);
 	}
 
 	@Override
