+++ src/main/java/de/kittybot/kittybot/commands/info/UptimeCommand.java
@@ -1,5 +1,7 @@
-package de.kittybot.kittybot.commands.info.info;
+package de.kittybot.kittybot.commands.info;
 
+import de.kittybot.kittybot.slashcommands.application.Category;
+import de.kittybot.kittybot.slashcommands.application.RunCommand;
 import de.kittybot.kittybot.slashcommands.application.options.SubCommand;
 import de.kittybot.kittybot.slashcommands.interaction.Interaction;
 import de.kittybot.kittybot.slashcommands.interaction.Options;
@@ -9,10 +11,10 @@ import de.kittybot.kittybot.utils.TimeUtils;
 import java.lang.management.ManagementFactory;
 
 @SuppressWarnings("unused")
-public class UptimeCommand extends SubCommand{
+public class UptimeCommand extends RunCommand{
 
 	public UptimeCommand(){
-		super("uptime", "Shows the bots uptime");
+		super("uptime", "Shows the bots uptime", Category.INFORMATION);
 	}
 
 	@Override
